TodoList/requirements.txt
Django 
django-filter 
djangorestframework 
Markdown 
djangorestframework_simplejwt


pip install -r .\requirements.txt  

python manage.py makemigrations 
python manage.py migrate 

python manage.py runserver